More sophisticated testing logic
Added a hard-coded default max test length value, so that if the
properties file parsed gives more than that many properties the user
gets a chance to choose a smaller value for this exam. The default max
is 50; this is b/c I don't currently want the prompt to show up when
operating against stateCaptials.props (which has 50 prompts).

The testing loop now has an outer loop that lets the user retest against
the entire set of prompts just tested against. This loop occurs after
the test size has already been chosen and the random prompts/properties
list has been selected.

So, for example, if run against the courtryCapitals.props file, which
has more than 50 prompts, the user gets a chance to choose a different
test size than the entire list. Maybe the user chooses 10. Then 10 of
the prompts are randomly selected and the user is tested against these.
Upon successful completion of this test the user can choose to retest
against this same set of 10 prompts or else exit.