Now printing out both promts and values each time
That is, each time a new round of a test takes place, the remaining
prompts and the remaining values are printed out. Currently they are
listed in the same order. Would probably be good to randomize the
orderings in the future to make it harder to "cheat" from this extra
detail.