Improved I/O using a BufferedReader, other changes
Attempting to use a BufferedReader on standard input rather than the
Scanner class of old. Initial trials of the changed class and tweaked
logic seem to look good.

Fixed the bug where integer input during the guessing phase was getting
ignored b/c it wasn't part of the regex of accepted characters. This is
independent from the switch to a BufferedReader.

Includes new and modified helper functions.

I/O could probably use to be improved further. Meep...