Finished removing references to states and capitals
Also changed 'prompt' to 'ioPrompt' in cases where its use was for I/O.
Turns out I'm using that term 'prompt' in two ways---as the key in the
mapping being tested (the test prompt) and as the output displayed when
I'm requesting user input (whether it be part of the test or other I/O
operations for configuration of the test. I should choose more
distinctive names for these. Probably.