Fixing 'keepOnlyAlphaNumeric' HelperFunctions logic
Wow. So it seems I initially meant that function to manually pick out
the characters I like, then decided to manaually pick out the characters
I /don't/ like, and then misunderstood the java.util.regex.Matcher
documentation and got the code quite wrong. I now use the
Matcher.replaceAll() function which is exactly what I should have used
in the first place.

Fwiw: My misunderstanding of the Matcher documentation was to think that
I could iterate through the individual matched sections using find() to
step to the first/next section and find the starting index of this
section with regionStart() to find the starting index of the current
section. I was wrong.