Fixing newline characters
For some reason, even though I'm using Linux/Vim, newlines were being
written as '\r\n' instead of just '\n'. This is a fix to that, and
hopefully Vim will be smart enough to keep using just '\n' now on.

My guess as to why this happened is that I started this project in
Eclipse/on Windows, and maybe Vim inspects the newline chars already
being used on a file to inform its choice on what to use going forward.
Best guess I have for now.